Stats on market UI for towers
