+++ Assets/Icons.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff62d9d20f4c3834ea0fbc9b3c62fdaf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
