+++ Assets/_Prefabs/Tower/Balista/Ballista 1.prefab
@@ -409,7 +409,7 @@ GameObject:
   - component: {fileID: 3565099394172681924}
   - component: {fileID: 3565099394172681923}
   m_Layer: 0
-  m_Name: Level 1
+  m_Name: Ballista 1
   m_TagString: balista_1
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -447,6 +447,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   radius: 5
   thumbnail: {fileID: 21300000, guid: c667c18da50f34eb49f5442da1886ee7, type: 3}
+  attack_descriptor: 0
+  speed_descriptor: 2
 --- !u!1 &3565099394211463872
 GameObject:
   m_ObjectHideFlags: 0
@@ -535,16 +537,6 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 3565099394172681924}
     m_Modifications:
-    - target: {fileID: 1660913033763128146, guid: b9c61004d54b72341ada963d2316caad,
-        type: 3}
-      propertyPath: m_Name
-      value: Range
-      objectReference: {fileID: 0}
-    - target: {fileID: 1660913033763128146, guid: b9c61004d54b72341ada963d2316caad,
-        type: 3}
-      propertyPath: m_IsActive
-      value: 1
-      objectReference: {fileID: 0}
     - target: {fileID: 1660913033763128157, guid: b9c61004d54b72341ada963d2316caad,
         type: 3}
       propertyPath: m_LocalPosition.x
@@ -600,6 +592,31 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 1660913033763128146, guid: b9c61004d54b72341ada963d2316caad,
+        type: 3}
+      propertyPath: m_Name
+      value: Range
+      objectReference: {fileID: 0}
+    - target: {fileID: 1660913033763128146, guid: b9c61004d54b72341ada963d2316caad,
+        type: 3}
+      propertyPath: m_IsActive
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1660913033763128156, guid: b9c61004d54b72341ada963d2316caad,
+        type: 3}
+      propertyPath: m_Enabled
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1660913033763128159, guid: b9c61004d54b72341ada963d2316caad,
+        type: 3}
+      propertyPath: m_Enabled
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1660913033763128158, guid: b9c61004d54b72341ada963d2316caad,
+        type: 3}
+      propertyPath: m_Mesh
+      value: 
+      objectReference: {fileID: 10206, guid: 0000000000000000e000000000000000, type: 0}
     - target: {fileID: 1660913033763128153, guid: b9c61004d54b72341ada963d2316caad,
         type: 3}
       propertyPath: onEnter2D.m_PersistentCalls.m_Calls.Array.size
@@ -664,21 +681,6 @@ PrefabInstance:
         type: 3}
       propertyPath: layerMask.m_Bits
       value: 256
-      objectReference: {fileID: 0}
-    - target: {fileID: 1660913033763128158, guid: b9c61004d54b72341ada963d2316caad,
-        type: 3}
-      propertyPath: m_Mesh
-      value: 
-      objectReference: {fileID: 10206, guid: 0000000000000000e000000000000000, type: 0}
-    - target: {fileID: 1660913033763128159, guid: b9c61004d54b72341ada963d2316caad,
-        type: 3}
-      propertyPath: m_Enabled
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 1660913033763128156, guid: b9c61004d54b72341ada963d2316caad,
-        type: 3}
-      propertyPath: m_Enabled
-      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents:
     - {fileID: 1660913033763128156, guid: b9c61004d54b72341ada963d2316caad, type: 3}
