+++ Assets/_Prefabs/Tower/Boomerang/Boomerang 1.prefab
@@ -128,6 +128,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   radius: 5
   thumbnail: {fileID: 21300000, guid: a1e069afc5922e242ab36239995c735e, type: 3}
+  attack_descriptor: 1
+  speed_descriptor: 1
 --- !u!1 &5992237024820890176
 GameObject:
   m_ObjectHideFlags: 0
