+++ Assets/_Prefabs/Tower/DamageSupport/Damage Support 1.prefab
@@ -48,6 +48,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   radius: 4
   thumbnail: {fileID: 21300000, guid: 09a9701945a80c844aa921b966db96db, type: 3}
+  attack_descriptor: 3
+  speed_descriptor: 3
   speedModifier: 1
   damageModifier: 1.5
 --- !u!1 &3565099394211463872
