+++ Assets/_Prefabs/UI/InGame_UI.prefab
@@ -46,6 +46,7 @@ MonoBehaviour:
   _maxWaves: 10
   _maxUnitsPerWave: 5
   unitsPerWaveIncrement: 5
+  waves: []
   onWaveStart:
     m_PersistentCalls:
       m_Calls:
@@ -361,16 +362,16 @@ MonoBehaviour:
   - {x: 15, y: 40, z: 45}
   tiers:
   - towers:
-    - {fileID: 3565099394172681925, guid: f7ab5ac4c677d334188c2bd168278963, type: 3}
-    - {fileID: 3565099394172681925, guid: 10fcb1531ea17224d960b64cf619d38a, type: 3}
+    - {fileID: 3565099394172681923, guid: f7ab5ac4c677d334188c2bd168278963, type: 3}
+    - {fileID: 3565099394172681923, guid: 10fcb1531ea17224d960b64cf619d38a, type: 3}
     price: 1
   - towers:
-    - {fileID: 3565099394172681925, guid: 3f808c6aa337c98458a20df965592c7a, type: 3}
-    - {fileID: 3565099394172681925, guid: 3f808c6aa337c98458a20df965592c7a, type: 3}
+    - {fileID: 5992237023386912483, guid: ebfb774d5a1216345b196c5bf81f4517, type: 3}
+    - {fileID: 6784485984623561150, guid: f7ba8f5a0f25a024db6997da940b0085, type: 3}
     price: 2
   - towers:
-    - {fileID: 3565099394172681925, guid: f7ba8f5a0f25a024db6997da940b0085, type: 3}
-    - {fileID: 3565099394172681925, guid: f7ba8f5a0f25a024db6997da940b0085, type: 3}
+    - {fileID: 6784485984623561150, guid: 00444b9d739411c4bac0bf5d512327ed, type: 3}
+    - {fileID: 3565099394172681923, guid: 3f808c6aa337c98458a20df965592c7a, type: 3}
     price: 3
 --- !u!1 &1194264112122302460
 GameObject:
@@ -2378,11 +2379,6 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 5767104169796231334}
     m_Modifications:
-    - target: {fileID: 7844059211303347297, guid: 342fd4b5ba64f0d428a330bd4b679334,
-        type: 3}
-      propertyPath: m_Name
-      value: Player
-      objectReference: {fileID: 0}
     - target: {fileID: 7844059211303347359, guid: 342fd4b5ba64f0d428a330bd4b679334,
         type: 3}
       propertyPath: onLevelUp.m_PersistentCalls.m_Calls.Array.size
@@ -2512,6 +2508,11 @@ PrefabInstance:
         type: 3}
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7844059211303347297, guid: 342fd4b5ba64f0d428a330bd4b679334,
+        type: 3}
+      propertyPath: m_Name
+      value: Player
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 342fd4b5ba64f0d428a330bd4b679334, type: 3}
