+++ Assets/_Scripts/UI/Market/MarketProduct.cs
@@ -6,13 +6,13 @@ using System;
 [Serializable]
 public struct MarketProduct {
 
-    public MarketProduct(GameObject product, int rarity, int price) {
+    public MarketProduct(Tower product, int rarity, int price) {
         this.product = product;
         this.rarity = rarity;
         this.price = price;
     }
 
-    public GameObject product;
+    public Tower product;
     public int        rarity;
     public int        price;
