+++ Assets/_Scripts/UI/Market/MarketSlot.cs
@@ -4,18 +4,28 @@ using UnityEngine.UI;
 
 public class MarketSlot : MonoBehaviour {
 
+
+
+    
     public                   MarketSlotEvent onTryBuying;
     [SerializeField] private Color[]         rarityColor;
     private                  MarketProduct   _product;
 
     private Text  _priceText;
     private Image _image;
-
+    private Text _attack;
+    private Text _speed;
+    [SerializeField] private GameObject _stats;
+    [SerializeField] private GameObject _support;
+    
 
     private void Awake() {
+        _support.SetActive(false);
+        _stats.SetActive(false);
         _priceText = transform.Find("Price").GetComponent<Text>();
         _image     = transform.Find("Image").GetComponent<Image>();
-
+        _attack = transform.Find("Stats").Find("Attack").Find("Value").GetComponent<Text>();
+        _speed = transform.Find("Stats").Find("Speed").Find("Value").GetComponent<Text>();
         // Register self click event
         GetComponent<Button>().onClick.AddListener(SelfClicked);
     }
@@ -39,9 +49,17 @@ public class MarketSlot : MonoBehaviour {
     public void SetProduct(MarketProduct product) {
         _product        = product;
         _priceText.text = _product.price.ToString();
-
+        _attack.text = product.product.attack_descriptor.ToString();
+        _speed.text = product.product.speed_descriptor.ToString();
+        if ( _attack.text == "Support" || _speed.text == "Support" ) {
+            _support.SetActive(true);
+            _stats.SetActive(false);
+        } else {
+            _stats.SetActive(true);
+            _support.SetActive(false);
+        }
         // Tower image
-        Tower tower = _product.product.GetComponent<Tower>();
+        Tower tower = _product.product;
         if ( tower != null && tower.thumbnail != null ) {
             _image.sprite = tower.thumbnail;
         }
